feat: use typescript and add node_module to path
