+++ src/babel.ts
@@ -1,0 +1,8 @@
+import { getBabelPlugins, Options } from './common';
+
+const babel = (config: any, options: Options) => ({
+  ...config,
+  plugins: [...config.plugins, ...getBabelPlugins(options)],
+});
+
+module.exports = babel;
